[ARITH] process comments
